convert size_t to uin32_t for predictable compilation on 32 bit and 64 bit platforms.

+++ app/libs/game.hpp
@@ -6,11 +6,11 @@
 #define DODDROID_GAME_HPP
 
 #include "error.hpp"
-#include <cstddef>
+#include <cstdint>
 
 AGE_RESULT game_init (struct android_app* p_app);
 AGE_RESULT game_process_motion_event_down (float x, float y);
-AGE_RESULT game_update (size_t delta_msecs);
+AGE_RESULT game_update (uint32_t delta_msecs);
 AGE_RESULT game_submit_present ();
 void game_shutdown ();
